+++ src/Products/PloneMeeting/migrations/migrate_to_4_1.py
@@ -1210,9 +1210,19 @@ class Migrate_To_4_1(Migrator):
                                  new_word="power_observer_type='powerobservers'")
         self.updateTALConditions(old_word='context.portal_plonemeeting',
                                  new_word='tool')
+        self.updateTALConditions(old_word='here.portal_plonemeeting',
+                                 new_word='tool')
         self.updateTALConditions(old_word='context.portal_plonemeeting.getMeetingConfig(context)',
                                  new_word='cfg')
+        self.updateTALConditions(old_word='here.portal_plonemeeting.getMeetingConfig(here)',
+                                 new_word='cfg')
+        self.updateTALConditions(old_word='context.portal_plonemeeting.getMeetingConfig(here)',
+                                 new_word='cfg')
+        self.updateTALConditions(old_word='here.portal_plonemeeting.getMeetingConfig(context)',
+                                 new_word='cfg')
         self.updateTALConditions(old_word='tool.getMeetingConfig(context)',
+                                 new_word='cfg')
+        self.updateTALConditions(old_word='tool.getMeetingConfig(here)',
                                  new_word='cfg')
         self._updateUsedAttributes()
         self._updateHistorizedAttributes()